Time: 2 ms (69.28%), Space: 5.9 MB (80.57%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     bool judgeSquareSum(int n) {
-    long int start = 0;
-    long int end = sqrt(n);
+    long long start = 0;
+    long long end = sqrt(n);
     while (start <= end)
     {
-        long int sum = start * start + end * end;
+        long long sum = start * start + end * end;
         if (sum == n)
             return true;
         else if (sum > n)
